Django/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
django-filter==24.2
Pillow==12.1.0
